fix: corrigir tipo de dado do campo expiresIn em LoginResponseDto

+++ src/main/java/com/example/springsecurity/dto/LoginResponseDto.java
@@ -1,5 +1,5 @@
 package com.example.springsecurity.dto;
 
-public record LoginResponseDto( String accessToken, Long expiresIn  ) {
-  
+public record LoginResponseDto(String accessToken, Integer expiresIn) {
+
 }
